fix: add target platform to build environment
Right now building with `--arch=arm` and then `--arch=arm64` will not
rebuild the `app.so`. This is because there's no reason to:
None of the input files changed on disk. The SDK maintains a content
hash per filepath. For arm and arm64, we're just using a different
`gen_snapshot` executable. So if the filepath changes, but the contents
of all files stay the same, the SDK won't rebuild that target.

This forces a different build environment for different target
platforms, which makes the flutter_tool unable to reuse the `app.so`
from a build from another target platform.